pyproject.toml
[tool.poetry]
name = "dj-ccr-news"
version = "0.1.0"
description = ""
authors = ["suband74 <suband@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
